fixation de l'affichage logo factorisation de navigateur en sous petite ligne d'instruction refactorisation de blog plus explicite
