--- src/navigation/Drawer.js
@@ -1,11 +1,0 @@
-import React from "react";
-
-function Drawer() {
-  return (
-    <div>
-      <h1>Drawer</h1>
-    </div>
-  );
-}
-
-export default Drawer;
