+++ src/pages/blogs/homeblogs.js
@@ -93,7 +93,9 @@ function Homeblogs({ search = "" }) {
             </Button>
           </div>
         </div>
-
+        {
+          // debut de data blog
+        }
         <div
           style={{
             marginTop: 25,
